Refactor trainee display components to Supervisor directory and update import paths
