+++ frontend/src/App.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import Create from './components/CreateTrainee';
-import Display from './components/DisplayTrainees';
+import Create from './components/Supervisor/CreateTrainee';
+import Display from './components/Supervisor/DisplayTrainees';
 const App=()=> {
 
   return (
